+++ src/reporting/__init__.py
@@ -1,7 +1,11 @@
-"""Public exports for reporting package."""
-
 from .errors import TemplateError
 from .template import Template, compile_template
+from .template_manifest import build_template_manifest, validate_template_context
 
-__all__ = ["Template", "TemplateError", "compile_template"]
-
+__all__ = [
+    "Template",
+    "TemplateError",
+    "compile_template",
+    "build_template_manifest",
+    "validate_template_context",
+]